+++ src/commands/createStaticWebApp/EnterAppLocationStep.ts
